+++ frontend/src/components/GroupView/MiddleBar/People/index.module.scss
@@ -20,7 +20,7 @@
 .AddMember div {
   width: 56px;
   border-radius: 50%;
-  border: 3px solid #bebebe;
+  border: 3px dashed #bebebe;
   text-align: center;
   margin-right: 1em;
 }
