+++ frontend/src/components/GroupView/MiddleBar/TaskQueue/index.module.css
@@ -1,0 +1,3 @@
+.TaskQueue {
+  height: 50vh;
+}
