+++ frontend/src/components/GroupView/MiddleBar/index.module.scss
@@ -1,0 +1,11 @@
+@import '../../../colors.scss';
+
+.MiddleBar {
+  margin: 0 1em;
+  background-color: $card-grey;
+  padding: 1em;
+  height: 100vh; 
+  color: $warm-grey;
+  display: flex;
+  flex-direction: column;
+}
