+++ frontend/src/components/GroupView/MiddleBar/index.tsx
@@ -1,6 +1,11 @@
 import React, { ReactElement } from 'react';
 import People from './People';
+import TaskQueue from './TaskQueue';
+import styles from './index.module.scss';
 
 export default (): ReactElement => (
-  <People groupMemberNames={['Darien Lopez', 'Sarah Johnson', 'Michelle Parker']} />
+  <div className={styles.MiddleBar}>
+    <TaskQueue />
+    <People groupMemberNames={['Darien Lopez', 'Sarah Johnson', 'Michelle Parker']} />
+  </div>
 );
