--- frontend/src/components/GroupView/index.module.scss
@@ -1,9 +1,0 @@
-@import '../../colors.scss';
-
-.MiddleBar {
-  margin: 0 1em;
-  background-color: $card-grey;
-  padding: 1em;
-  height: 100vh; 
-  color: $warm-grey;
-}
