+++ frontend/src/components/SideBar/index.module.scss
@@ -1,7 +1,7 @@
 @import '../../colors.scss';
 
 .SideBar {
-  width: 5%;
+  width: 60px;
   display: flex;
   flex-direction: column;
 }
